Moved extension to separate file.
