+++ I Got Gas/Extensions.swift
@@ -1,0 +1,22 @@
+//
+//  Extensions.swift
+//  I Got Gas
+//
+//  Created by Isaac Lyons on 7/30/20.
+//  Copyright © 2020 Blizzard Skeleton. All rights reserved.
+//
+
+import Foundation
+
+extension NumberFormatter {
+    static var decimal: NumberFormatter {
+        let formatter = NumberFormatter()
+        formatter.numberStyle = .decimal
+        return formatter
+    }
+    static var currency: NumberFormatter {
+        let formatter = NumberFormatter()
+        formatter.numberStyle = .currency
+        return formatter
+    }
+}
